Modify Makefile

+++ Makefile
@@ -1,30 +1,38 @@
+BIN=$$(pwd)/bin
+TMP=$$(pwd)/tmp
+GOBIN=$(if $$(go env GOBIN),$$(go env GOPATH)/bin,$$(go env GOBIN))
+
 setup:
-	mkdir tmp_migrate
-	curl -sSOL https://github.com/golang-migrate/migrate/releases/download/v4.15.0/migrate.linux-amd64.tar.gz
-	mv *tar.gz tmp_migrate && cd tmp_migrate && tar -xvzf *.tar.gz
-	sudo mv tmp_migrate/migrate /usr/local/bin
-	rm -r tmp_migrate
-	curl -sSLO "https://github.com/bufbuild/buf/releases/download/v1.0.0-rc6/buf-$$(uname -s)-$$(uname -m)"
-	chmod 755 buf-$$(uname -s)-$$(uname -m)
-	sudo mv ./buf-$$(uname -s)-$$(uname -m)  /usr/local/bin/buf
-	go install google.golang.org/protobuf/cmd/protoc-gen-go@v1.27.1
-	go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.1.0
-	GO111MODULE=on GOBIN=$$(go env GOPATH)/bin go install github.com/uber/prototool/cmd/prototool@dev
-	mkdir tmp_protoc
-	curl -sSLO "https://github.com/protocolbuffers/protobuf/releases/download/v3.19.1/protoc-3.19.1-$$(uname -s)-$$(uname -m).zip"
-	mv *.zip tmp_protoc && cd tmp_protoc &&	unzip protoc-3.19.1-$$(uname -s)-$$(uname -m).zip
-	sudo mv ./tmp_protoc/include /usr/local/bin
-	sudo mv ./tmp_protoc/bin/protoc /usr/local/bin
-	rm -rf tmp_protoc
+	if [ ! -e $(BIN)/migrate ]; then\
+		if [ ! -d $(TMP) ]; then mkdir $(TMP); fi;\
+		curl -sSL https://github.com/golang-migrate/migrate/releases/download/v4.15.0/migrate.linux-amd64.tar.gz -o $(TMP)/migrate.tar.gz;\
+		tar -xvzf $(TMP)/migrate.tar.gz -C $(TMP);\
+		mv $(TMP)/migrate $(BIN);\
+		rm -rf $(TMP);\
+	fi
+	if [ ! -e $(BIN)/buf ]; then\
+		curl -sSL "https://github.com/bufbuild/buf/releases/download/v1.0.0-rc6/buf-$$(uname -s)-$$(uname -m)" -o $(BIN)/buf;\
+		chmod 755 $(BIN)/buf;\
+	fi
+	if [ ! -e $(GOBIN)/proto-gen-go ]; then	go install google.golang.org/protobuf/cmd/protoc-gen-go@v1.27.1; fi
+	if [ ! -e $(GOBIN)/protoc-gen-go-grpc ]; then go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@v1.1.0; fi
+	if [ ! -e $(GOBIN)/prototool ]; then GO111MODULE=on GOBIN=$(GOBIN) go install github.com/uber/prototool/cmd/prototool@dev; fi
+	if [ ! -d $(BIN)/include ]; then\
+		if [ ! -d $(TMP) ]; then mkdir $(TMP); fi;\
+		curl -sSL "https://github.com/protocolbuffers/protobuf/releases/download/v3.19.1/protoc-3.19.1-$$(uname -s)-$$(uname -m).zip" -o $(TMP)/protoc.zip;\
+		unzip $(TMP)/protoc.zip -d $(TMP);\
+		mv $(TMP)/include $(BIN);\
+		rm -rf $(TMP);\
+	fi
 
 migrate-up:
-	migrate -source file://migrations -database mysql://localhost:3306 up
+	$(BIN)/migrate -source file://migrations -database mysql://localhost:3306 up
 
 run:
-	./bin/up.sh
+	$(BIN)/up.sh
 
 lint:
-	buf lint
+	$(BIN)/buf lint
 
 gen:
 	protoc -I=./proto --go_out=./app/proto --go_opt=paths=source_relative --go-grpc_out=require_unimplemented_servers=false:./app/proto --go-grpc_opt=paths=source_relative $$(find ./proto/app/services/ -name "*.proto")
